frosttss: Go-compatible Key + SignatureData
Port frosttss/key.go and signature.go. Key persists to the exact Go JSON
shape — capitalized field names, big.Int values as bare decimal numbers,
points as {"Curve":"ed25519","Coords":[X,Y]} (affine, via purecrypto 0.5.1
to_affine) — so saved shares interoperate across both libraries. Includes
validate_basic (share/commitment binding), subset_for_parties, zeroize,
and a typed frosttss::Error. SignatureData mirrors the Go output (R/S/
Signature/M, base64 bytes).

Adds shared serde helpers: tss::bigint (big.Int<->decimal JSON number) and
tss::b64 (Go []byte<->base64). 41 tests incl. JSON-shape and legacy
(no-ChainCode) parsing.

+++ src/tss/bigint.rs
@@ -86,7 +86,7 @@ pub fn decimal_to_be(s: &str) -> Result<Vec<u8>, DecimalError> {
     // Accumulate big-endian base-256: bytes = bytes * 10 + digit.
     let mut bytes: Vec<u8> = vec![0];
     for ch in s.chars() {
-        let digit = ch.to_digit(10).ok_or(DecimalError("non-digit character"))? as u32;
+        let digit = ch.to_digit(10).ok_or(DecimalError("non-digit character"))?;
         let mut carry = digit;
         for b in bytes.iter_mut().rev() {
             let v = (*b as u32) * 10 + carry;
